Start on Users saving their workspaces.
Users can now (unsafely) save their workspaces. We added a field to
the mongoose User schema called expertWorkspaces which is an array
that will store our workspace objects. This saving is unsafe because
we are only pushing the workspace objects in and not allowing
for overwriting. We're also not properly checking for failed writes.